RSS/requirements.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
celery==3.1.26.post2
certifi==2018.4.16
chardet==3.0.4
cycler==0.10.0
Django==2.0.7
django-celery==3.2.2
djangorestframework==3.8.2
docutils==0.14
gunicorn==19.8.1
idna==2.6
Kivy==1.10.0
Kivy-examples==1.10.0
Kivy-Garden==0.1.4
kivy.deps.angle==0.1.6
kivy.deps.glew==0.1.9
kivy.deps.gstreamer==0.1.12
kivy.deps.sdl2==0.1.17
kombu==3.0.37
matplotlib==2.1.1
numpy==1.14.0
Pillow==5.1.0
psycopg2==2.7.5
pygame==1.9.3
Pygments==2.2.0
pyparsing==2.2.0
pypiwin32==223
python-dateutil==2.6.1
pytz==2017.3
pywin32==223
redis==2.10.3
requests==2.18.4
scipy==1.0.0
six==1.11.0
urllib3==1.22
vine==1.1.4
